refactor(orchestrator): extract state/progress dataclasses to a models module
Per the architecture audit: the orchestrator should read as control flow,
not data definitions. Moved the durable state types (LineageEntry,
AbandonedEntry, UnitTrack) and the display snapshot (OrchestratorProgress),
plus the unit_id helper, into qallm.orchestrator_models.

* orchestrator.py: 851 -> ~615 lines, now focused on __init__, run(), the
  baseline/round loop, and summary building. Imports the models back with
  their original names (and unit_id as _unit_id) so nothing else changes.
* orchestrator_models.py: the extracted dataclasses, importable directly
  and re-exported from orchestrator for back-compat.
* Removed 133 now-unused imports across the new api router modules (ruff).

No behaviour change. Suite: 376 passed; ruff clean on the changed files.
