+++ src/qallm/api/routers/verification.py
@@ -7,31 +7,16 @@ file boundary moved.
 
 from __future__ import annotations
 
-import json
 import logging
-import os
-import shutil
-import tempfile
-import uuid
-import zipfile
-from datetime import datetime
-from pathlib import Path
-from typing import List, Optional
 
-from fastapi import APIRouter, File, Form, HTTPException, UploadFile
-from fastapi.responses import FileResponse
+from fastapi import APIRouter, HTTPException
 
 from qallm.api.core import (
-    MODEL_CATALOG,
-    check_model_available,
     get_state,
-    model_label,
-    parse_model_id,
     sessions,
 )
-from qallm.config import settings
-from qallm.jobs import JobConflictError, JobStatus, get_store
-from qallm.orchestrator import QALLMOrchestrator, LLM_PROVIDERS
+from qallm.jobs import JobConflictError, get_store
+from qallm.orchestrator import QALLMOrchestrator
 
 logger = logging.getLogger(__name__)
